Fixed errors in the StudentManager.readFromFile method.
The while condition was changed from the return value of hasNextLine to the return value of hasNext due to hasNextLine causing an error when the last line is blank. Also added a second fileIn.next() to the name field because code was only reading the first name and then attempting to read the last name as a float.